Split prj to isolate layers
