+++ .idea/encodings.xml
@@ -1,6 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/API/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/API/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Application/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/Infra/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
   </component>
