+++ Application/src/main/java/com/phonebook/domain/DeviceDetails.java
