+++ Application/src/main/java/com/phonebook/exceptions/DeviceAlreadyInUseException.java
