+++ Application/src/main/java/com/phonebook/exceptions/DeviceNotFoundException.java
