+++ Application/src/main/java/com/phonebook/exceptions/InternalException.java
