+++ Application/src/main/java/com/phonebook/exceptions/NotAuthorizedException.java
