+++ Application/src/main/java/com/phonebook/mappers/DevicesMapper.java
