+++ Infra/src/main/java/com/phonebook/model/Device.java
